+++ Kooboo.Sites/Scripting/Global/Database/kDatabase.cs
@@ -6,31 +6,27 @@ using System.ComponentModel;
 
 namespace KScript
 {
-    [KValueType(typeof(KTable))]
-    public class kDatabase
+    public class kDatabase : IDatabase
     {
         private RenderContext context { get; set; }
         public kDatabase(RenderContext context)
         {
             this.context = context;
         }
-
-        [Description("Return the kScript database table object, if the table is not exists, it will be created.")]
-        public KTable GetTable(string Name)
+       
+        public ITable GetTable(string Name)
         {
             var db = Kooboo.Data.DB.GetKDatabase(this.context.WebSite);
             var tb = Kooboo.Data.DB.GetOrCreateTable(db, Name);
             return new KTable(tb, this.context);
         }
 
-        [KIgnore]
-        public KTable Table(string Name)
+        public ITable Table(string Name)
         {
             return GetTable(Name);
         }
 
-        [KIgnore]
-        public KTable this[string key]
+        public ITable this[string key]
         {
             get
             {
@@ -41,6 +37,6 @@ namespace KScript
 
             }
         }
-        
+
     }
 }
