+++ Kooboo.Sites/Scripting/Global/Sqlite/RelationModel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Kooboo.Sites.Scripting.Global.Sqlite
+{
+    public class RelationModel
+    {
+        public string Table { get; set; }
+        public string From { get; set; }
+        public string To { get; set; }
+    }
+}
