+++ Kooboo.Sites/Scripting/Interfaces/ISqliteDatabase.cs
@@ -1,0 +1,8 @@
+﻿namespace KScript
+{
+    public interface ISqliteDatabase : IDatabase
+    {
+        int Execute(string sql);
+        object[] Query(string sql);
+    }
+}
